refactor common utils
