+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark/src/main/java/org/apache/pinot/plugin/ingestion/batch/spark/SparkSegmentGenerationJobRunner.java
@@ -18,16 +18,10 @@
  */
 package org.apache.pinot.plugin.ingestion.batch.spark;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Preconditions;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.Serializable;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
 import java.nio.file.PathMatcher;
 import java.nio.file.Paths;
@@ -37,11 +31,9 @@ import java.util.List;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.MapConfiguration;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
-import org.apache.pinot.common.config.TableConfig;
 import org.apache.pinot.common.utils.TarGzCompressionUtils;
 import org.apache.pinot.plugin.ingestion.batch.common.SegmentGenerationTaskRunner;
-import org.apache.pinot.spi.data.Schema;
+import org.apache.pinot.plugin.ingestion.batch.common.SegmentGenerationUtils;
 import org.apache.pinot.spi.filesystem.PinotFS;
 import org.apache.pinot.spi.filesystem.PinotFSFactory;
 import org.apache.pinot.spi.ingestion.batch.runner.IngestionJobRunner;
@@ -61,7 +53,6 @@ import org.slf4j.LoggerFactory;
 public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Serializable {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(SparkSegmentGenerationJobRunner.class);
-  private static final String OFFLINE = "OFFLINE";
   private static final String DEPS_JAR_DIR = "dependencyJarDir";
   private static final String STAGING_DIR = "stagingDir";
 
@@ -72,34 +63,6 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
 
   public SparkSegmentGenerationJobRunner(SegmentGenerationJobSpec spec) {
     init(spec);
-  }
-
-  private static String generateSchemaURI(String controllerUri, String table) {
-    return String.format("%s/tables/%s/schema", controllerUri, table);
-  }
-
-  private static String generateTableConfigURI(String controllerUri, String table) {
-    return String.format("%s/tables/%s", controllerUri, table);
-  }
-
-  /**
-   * Generate a relative output directory path when `useRelativePath` flag is on.
-   * This method will compute the relative path based on `inputFile` and `baseInputDir`,
-   * then apply only the directory part of relative path to `outputDir`.
-   * E.g.
-   *    baseInputDir = "/path/to/input"
-   *    inputFile = "/path/to/input/a/b/c/d.avro"
-   *    outputDir = "/path/to/output"
-   *    getRelativeOutputPath(baseInputDir, inputFile, outputDir) = /path/to/output/a/b/c
-   */
-  public static URI getRelativeOutputPath(URI baseInputDir, URI inputFile, URI outputDir) {
-    URI relativePath = baseInputDir.relativize(inputFile);
-    Preconditions.checkState(relativePath.getPath().length() > 0 && !relativePath.equals(inputFile),
-        "Unable to extract out the relative path based on base input path: " + baseInputDir);
-    String outputDirStr = outputDir.toString();
-    outputDir = !outputDirStr.endsWith("/") ? URI.create(outputDirStr.concat("/")) : outputDir;
-    URI relativeOutputURI = outputDir.resolve(relativePath).resolve(".");
-    return relativeOutputURI;
   }
 
   @Override
@@ -125,7 +88,8 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
         throw new RuntimeException("Missing property 'schemaURI' in 'tableSpec'");
       }
       PinotClusterSpec pinotClusterSpec = _spec.getPinotClusterSpecs()[0];
-      String schemaURI = generateSchemaURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
+      String schemaURI = SegmentGenerationUtils
+          .generateSchemaURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
       _spec.getTableSpec().setSchemaURI(schemaURI);
     }
     if (_spec.getTableSpec().getTableConfigURI() == null) {
@@ -133,8 +97,8 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
         throw new RuntimeException("Missing property 'tableConfigURI' in 'tableSpec'");
       }
       PinotClusterSpec pinotClusterSpec = _spec.getPinotClusterSpecs()[0];
-      String tableConfigURI =
-          generateTableConfigURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
+      String tableConfigURI = SegmentGenerationUtils
+          .generateTableConfigURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
       _spec.getTableSpec().setTableConfigURI(tableConfigURI);
     }
     if (_spec.getExecutionFrameworkSpec().getExtraConfigs() == null) {
@@ -252,8 +216,9 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
         taskSpec.setInputFilePath(localInputDataFile.getAbsolutePath());
         taskSpec.setOutputDirectoryPath(localOutputTempDir.getAbsolutePath());
         taskSpec.setRecordReaderSpec(_spec.getRecordReaderSpec());
-        taskSpec.setSchema(getSchema());
-        taskSpec.setTableConfig(getTableConfig().toJsonNode());
+        taskSpec.setSchema(SegmentGenerationUtils.getSchema(_spec.getTableSpec().getSchemaURI()));
+        taskSpec.setTableConfig(
+            SegmentGenerationUtils.getTableConfig(_spec.getTableSpec().getTableConfigURI()).toJsonNode());
         taskSpec.setSequenceId(idx);
         taskSpec.setSegmentNameGeneratorSpec(_spec.getSegmentNameGeneratorSpec());
 
@@ -272,7 +237,8 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
             DataSize.fromBytes(uncompressedSegmentSize), DataSize.fromBytes(compressedSegmentSize));
         //move segment to output PinotFS
         URI outputSegmentTarURI =
-            getRelativeOutputPath(finalInputDirURI, inputFileURI, finalOutputDirURI).resolve(segmentTarFileName);
+            SegmentGenerationUtils.getRelativeOutputPath(finalInputDirURI, inputFileURI, finalOutputDirURI)
+                .resolve(segmentTarFileName);
         LOGGER.info("Trying to move segment tar file from: [{}] to [{}]", localSegmentTarFile, outputSegmentTarURI);
         if (!_spec.isOverwriteOutput() && PinotFSFactory.create(outputSegmentTarURI.getScheme())
             .exists(outputSegmentTarURI)) {
@@ -294,87 +260,6 @@ public class SparkSegmentGenerationJobRunner implements IngestionJobRunner, Seri
         LOGGER.info("Trying to clean up staging directory: [{}]", stagingDirURI);
         outputDirFS.delete(stagingDirURI, true);
       }
-    }
-  }
-
-  private Schema getSchema() {
-    URI schemaURI;
-    try {
-      schemaURI = new URI(_spec.getTableSpec().getSchemaURI());
-    } catch (URISyntaxException e) {
-      throw new RuntimeException("Schema URI is not valid - '" + _spec.getTableSpec().getSchemaURI() + "'", e);
-    }
-    String scheme = schemaURI.getScheme();
-    String schemaJson;
-    if (PinotFSFactory.isSchemeSupported(scheme)) {
-      // Try to use PinotFS to read schema URI
-      PinotFS pinotFS = PinotFSFactory.create(scheme);
-      InputStream schemaStream;
-      try {
-        schemaStream = pinotFS.open(schemaURI);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to fetch schema from PinotFS - '" + schemaURI + "'", e);
-      }
-      try {
-        schemaJson = IOUtils.toString(schemaStream, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to read from schema file data stream on Pinot fs - '" + schemaURI + "'", e);
-      }
-    } else {
-      // Try to directly read from URI.
-      try {
-        schemaJson = IOUtils.toString(schemaURI, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to read from Schema URI - '" + schemaURI + "'", e);
-      }
-    }
-    try {
-      return Schema.fromString(schemaJson);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode Pinot schema from json string - '" + schemaJson + "'", e);
-    }
-  }
-
-  private TableConfig getTableConfig() {
-    URI tableConfigURI;
-    try {
-      tableConfigURI = new URI(_spec.getTableSpec().getTableConfigURI());
-    } catch (URISyntaxException e) {
-      throw new RuntimeException("Table config URI is not valid - '" + _spec.getTableSpec().getTableConfigURI() + "'",
-          e);
-    }
-    String scheme = tableConfigURI.getScheme();
-    String tableConfigJson;
-    if (PinotFSFactory.isSchemeSupported(scheme)) {
-      // Try to use PinotFS to read table config URI
-      PinotFS pinotFS = PinotFSFactory.create(scheme);
-      try {
-        tableConfigJson = IOUtils.toString(pinotFS.open(tableConfigURI), StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to open table config file stream on Pinot fs - '" + tableConfigURI + "'", e);
-      }
-    } else {
-      try {
-        tableConfigJson = IOUtils.toString(tableConfigURI, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException(
-            "Failed to read from table config file data stream on Pinot fs - '" + tableConfigURI + "'", e);
-      }
-    }
-    // Controller API returns a wrapper of table config.
-    JsonNode tableJsonNode;
-    try {
-      tableJsonNode = new ObjectMapper().readTree(tableConfigJson);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode table config into JSON from String - '" + tableConfigJson + "'", e);
-    }
-    if (tableJsonNode.has(OFFLINE)) {
-      tableJsonNode = tableJsonNode.get(OFFLINE);
-    }
-    try {
-      return TableConfig.fromJsonConfig(tableJsonNode);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode table config from JSON - '" + tableJsonNode + "'", e);
     }
   }
