+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-standalone/src/main/java/org/apache/pinot/plugin/ingestion/batch/standalone/SegmentGenerationJobRunner.java
@@ -18,15 +18,8 @@
  */
 package org.apache.pinot.plugin.ingestion.batch.standalone;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Preconditions;
 import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
 import java.nio.file.PathMatcher;
 import java.nio.file.Paths;
@@ -35,10 +28,10 @@ import java.util.List;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.MapConfiguration;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
 import org.apache.pinot.common.config.TableConfig;
 import org.apache.pinot.common.utils.TarGzCompressionUtils;
 import org.apache.pinot.plugin.ingestion.batch.common.SegmentGenerationTaskRunner;
+import org.apache.pinot.plugin.ingestion.batch.common.SegmentGenerationUtils;
 import org.apache.pinot.spi.data.Schema;
 import org.apache.pinot.spi.filesystem.PinotFS;
 import org.apache.pinot.spi.filesystem.PinotFSFactory;
@@ -56,7 +49,6 @@ import org.slf4j.LoggerFactory;
 public class SegmentGenerationJobRunner implements IngestionJobRunner {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(SegmentGenerationJobRunner.class);
-  private static final String OFFLINE = "OFFLINE";
 
   private SegmentGenerationJobSpec _spec;
 
@@ -65,34 +57,6 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
 
   public SegmentGenerationJobRunner(SegmentGenerationJobSpec spec) {
     init(spec);
-  }
-
-  private static String generateSchemaURI(String controllerUri, String table) {
-    return String.format("%s/tables/%s/schema", controllerUri, table);
-  }
-
-  private static String generateTableConfigURI(String controllerUri, String table) {
-    return String.format("%s/tables/%s", controllerUri, table);
-  }
-
-  /**
-   * Generate a relative output directory path when `useRelativePath` flag is on.
-   * This method will compute the relative path based on `inputFile` and `baseInputDir`,
-   * then apply only the directory part of relative path to `outputDir`.
-   * E.g.
-   *    baseInputDir = "/path/to/input"
-   *    inputFile = "/path/to/input/a/b/c/d.avro"
-   *    outputDir = "/path/to/output"
-   *    getRelativeOutputPath(baseInputDir, inputFile, outputDir) = /path/to/output/a/b/c
-   */
-  public static URI getRelativeOutputPath(URI baseInputDir, URI inputFile, URI outputDir) {
-    URI relativePath = baseInputDir.relativize(inputFile);
-    Preconditions.checkState(relativePath.getPath().length() > 0 && !relativePath.equals(inputFile),
-        "Unable to extract out the relative path based on base input path: " + baseInputDir);
-    String outputDirStr = outputDir.toString();
-    outputDir = !outputDirStr.endsWith("/") ? URI.create(outputDirStr.concat("/")) : outputDir;
-    URI relativeOutputURI = outputDir.resolve(relativePath).resolve(".");
-    return relativeOutputURI;
   }
 
   @Override
@@ -118,7 +82,8 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
         throw new RuntimeException("Missing property 'schemaURI' in 'tableSpec'");
       }
       PinotClusterSpec pinotClusterSpec = _spec.getPinotClusterSpecs()[0];
-      String schemaURI = generateSchemaURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
+      String schemaURI = SegmentGenerationUtils
+          .generateSchemaURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
       _spec.getTableSpec().setSchemaURI(schemaURI);
     }
     if (_spec.getTableSpec().getTableConfigURI() == null) {
@@ -126,8 +91,8 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
         throw new RuntimeException("Missing property 'tableConfigURI' in 'tableSpec'");
       }
       PinotClusterSpec pinotClusterSpec = _spec.getPinotClusterSpecs()[0];
-      String tableConfigURI =
-          generateTableConfigURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
+      String tableConfigURI = SegmentGenerationUtils
+          .generateTableConfigURI(pinotClusterSpec.getControllerURI(), _spec.getTableSpec().getTableName());
       _spec.getTableSpec().setTableConfigURI(tableConfigURI);
     }
   }
@@ -196,8 +161,8 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
       FileUtils.forceMkdir(localOutputTempDir);
 
       //Read TableConfig, Schema
-      Schema schema = getSchema();
-      TableConfig tableConfig = getTableConfig();
+      Schema schema = SegmentGenerationUtils.getSchema(_spec.getTableSpec().getSchemaURI());
+      TableConfig tableConfig = SegmentGenerationUtils.getTableConfig(_spec.getTableSpec().getTableConfigURI());
 
       //iterate on the file list, for each
       for (int i = 0; i < filteredFiles.size(); i++) {
@@ -235,8 +200,8 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
         LOGGER.info("Size for segment: {}, uncompressed: {}, compressed: {}", segmentName,
             DataSize.fromBytes(uncompressedSegmentSize), DataSize.fromBytes(compressedSegmentSize));
         //move segment to output PinotFS
-        URI outputSegmentTarURI =
-            getRelativeOutputPath(inputDirURI, inputFileURI, outputDirURI).resolve(segmentTarFileName);
+        URI outputSegmentTarURI = SegmentGenerationUtils.getRelativeOutputPath(inputDirURI, inputFileURI, outputDirURI)
+            .resolve(segmentTarFileName);
         if (!_spec.isOverwriteOutput() && outputDirFS.exists(outputSegmentTarURI)) {
           LOGGER.warn("Not overwrite existing output segment tar file: {}", outputDirFS.exists(outputSegmentTarURI));
         } else {
@@ -249,87 +214,6 @@ public class SegmentGenerationJobRunner implements IngestionJobRunner {
     } finally {
       //clean up
       FileUtils.deleteDirectory(localTempDir);
-    }
-  }
-
-  private Schema getSchema() {
-    URI schemaURI;
-    try {
-      schemaURI = new URI(_spec.getTableSpec().getSchemaURI());
-    } catch (URISyntaxException e) {
-      throw new RuntimeException("Schema URI is not valid - '" + _spec.getTableSpec().getSchemaURI() + "'", e);
-    }
-    String scheme = schemaURI.getScheme();
-    String schemaJson;
-    if (PinotFSFactory.isSchemeSupported(scheme)) {
-      // Try to use PinotFS to read schema URI
-      PinotFS pinotFS = PinotFSFactory.create(scheme);
-      InputStream schemaStream;
-      try {
-        schemaStream = pinotFS.open(schemaURI);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to fetch schema from PinotFS - '" + schemaURI + "'", e);
-      }
-      try {
-        schemaJson = IOUtils.toString(schemaStream, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to read from schema file data stream on Pinot fs - '" + schemaURI + "'", e);
-      }
-    } else {
-      // Try to directly read from URI.
-      try {
-        schemaJson = IOUtils.toString(schemaURI, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to read from Schema URI - '" + schemaURI + "'", e);
-      }
-    }
-    try {
-      return Schema.fromString(schemaJson);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode Pinot schema from json string - '" + schemaJson + "'", e);
-    }
-  }
-
-  private TableConfig getTableConfig() {
-    URI tableConfigURI;
-    try {
-      tableConfigURI = new URI(_spec.getTableSpec().getTableConfigURI());
-    } catch (URISyntaxException e) {
-      throw new RuntimeException("Table config URI is not valid - '" + _spec.getTableSpec().getTableConfigURI() + "'",
-          e);
-    }
-    String scheme = tableConfigURI.getScheme();
-    String tableConfigJson;
-    if (PinotFSFactory.isSchemeSupported(scheme)) {
-      // Try to use PinotFS to read table config URI
-      PinotFS pinotFS = PinotFSFactory.create(scheme);
-      try {
-        tableConfigJson = IOUtils.toString(pinotFS.open(tableConfigURI), StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException("Failed to open table config file stream on Pinot fs - '" + tableConfigURI + "'", e);
-      }
-    } else {
-      try {
-        tableConfigJson = IOUtils.toString(tableConfigURI, StandardCharsets.UTF_8);
-      } catch (IOException e) {
-        throw new RuntimeException(
-            "Failed to read from table config file data stream on Pinot fs - '" + tableConfigURI + "'", e);
-      }
-    }
-    // Controller API returns a wrapper of table config.
-    JsonNode tableJsonNode;
-    try {
-      tableJsonNode = new ObjectMapper().readTree(tableConfigJson);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode table config into JSON from String - '" + tableConfigJson + "'", e);
-    }
-    if (tableJsonNode.has(OFFLINE)) {
-      tableJsonNode = tableJsonNode.get(OFFLINE);
-    }
-    try {
-      return TableConfig.fromJsonConfig(tableJsonNode);
-    } catch (IOException e) {
-      throw new RuntimeException("Failed to decode table config from JSON - '" + tableJsonNode + "'", e);
     }
   }
 }
